Refactoring; Removing distortion; Computing loss
The codebase is refactored for Python 3.x. Compatibility with Python 2.x
is being kept.
You can specify distortion parameters in the configuration file and it
will remove the distortion on the fly.
At the end of the calibration a reprojection error will be printed and
saved to the calibration file.
Updated README.